fix: move native module listener logic out of constructor.
1. On ios, eagerly checking `instanceof` in constructor
makes all modules to load on app launch, breaking the logic for
lazy loading. Secondly I don't see any benefit of doing it in the
contructor, so I have made the normal class instance methods
insteads which access the nativeModule when called only.

2. On android, I have added polyfill methods that do nothing actually
but only polyfill the methods on iOS to prevent unintended crashes